pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "synology-mcp"
version = "0.1.0"
description = "MCP server for Synology NAS — manage files, containers, and more via Claude"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Chris Means" },
]
keywords = ["mcp", "synology", "nas", "file-station", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "mcp>=1.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "keyring>=25.0",
    "pydantic>=2.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "respx>=0.22",
]

[project.scripts]
synology-mcp = "synology_mcp.cli:main"

[project.urls]
Homepage = "https://github.com/cmeans/synology-mcp"
Repository = "https://github.com/cmeans/synology-mcp"
Issues = "https://github.com/cmeans/synology-mcp/issues"
Documentation = "https://github.com/cmeans/synology-mcp/tree/main/docs"

[tool.hatch.build.targets.wheel]
packages = ["src/synology_mcp"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = ["integration: requires a real Synology NAS (not run in CI)"]
addopts = "-m 'not integration'"
